replaced modal component for macro screen
